domparsing/domparsing_xmlserializer_toString.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<html>
 <head>
  <title>domparsing Test: XMLSerializer.serializeToString</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <!-- domparsing_XMLSerializer_serializeToStrin -->
  <h1>domparsing_XMLSerializer_serializeToString</h1>
  <div id="log"></div>
  <script>
    function createXmlDoc(){
      var input = '<?xml version="1.0" encoding="UTF-8"?><root><child1>value1</child1></root>',
      parser = new DOMParser();
      doc = parser.parseFromString(input, 'text/xml');
      return doc;
    }

    test(function() {
      var serializer = new XMLSerializer (),
          root = createXmlDoc().documentElement;
      //Please clean up the console before submitting the code
      //alert(root);
      xmlString = serializer.serializeToString(root);
           
      assert_equals(xmlString, "<root><child1>value1</child1></root>");
    }, 'check whether a require well-formed xml document could be parsed into DomString properly');
  </script>
</body>
</html>
